% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_means.R
\name{od_means}
\alias{od_means}
\title{ODMeans Function}
\usage{
od_means(
  data,
  numKGlobal,
  limitSeparationGlobal,
  maxDistGlobal,
  distHierarchical,
  numKLocal,
  limitSeparationLocal,
  maxDistLocal
)
}
\arguments{
\item{data}{A data frame with four columns:\cr
Initial Latitude | Initial Longitude | Final Latitude | Final Longitude}

\item{numKGlobal}{Initial number of clusters in the first call of k-means in the global hierarchy.}

\item{limitSeparationGlobal}{Within cluster distance threshold to determine if a global cluster must be separated into two new clusters.}

\item{maxDistGlobal}{Meter distance threshold used to re-estimate centroids in global hierarchy.}

\item{distHierarchical}{Meter distance threshold between origin and destination to generate new local clusters from a first layer cluster}

\item{numKLocal}{Initial number of clusters in the first call of k-means in the local hierarchy.}

\item{limitSeparationLocal}{Within cluster distance threshold to determine if a local cluster must be separated into two new clusters.}

\item{maxDistLocal}{Meter distance threshold used to re-estimate centroids in local hierarchy.}
}
\value{
Returns an S3 class object similar to kmeans S3 Class, with eight properties.
}
\description{
ODMeans Function
}
\examples{
data(ODMeansTaxiData)
odmeans_data = od_means(ODMeansTaxiData, 10, 30, 1000, 2200, 3, 5, 100)
}
