% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODMeansTaxiData.R
\docType{data}
\name{ODMeansTaxiData}
\alias{ODMeansTaxiData}
\title{Origin-Destination Taxi data}
\format{
A data frame with 452166 rows and 4 variables:
\describe{
  \item{OriginLatitude}{Consists of the origin latitude dimension}
  \item{OriginLongitude}{Consists of the origin longitude dimension}
  \item{DestinationLatitude}{Consists of the destination latitude dimension}
  \item{DestinationLongitude}{Consists of the destination longitude dimension}
  ...
}
}
\source{
Taxi data
}
\usage{
ODMeansTaxiData
}
\description{
A dataset containing 1700 Origin-Destination points
}
\keyword{datasets}
